[be] feat: GamerStatus, GamerLog 관련 클래스들 생성

+++ backend/src/main/java/com/ssafy/be/gamer/model/domain/GamerLog.java
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 @ToString
 public class GamerLog {
     @Id
-    private int gamerLogId;
+    private Long gamerLogId;
     private int gamerId;
     private int gameId;
     private int teamId;
